Fix batch_label type

+++ classification/client_classification.py
@@ -60,7 +60,7 @@ def train_single_label(model, train_loader, criterion, optimizer, classes_names,
 
         for batch_idx, (images, batch_labels) in enumerate(train_loader):
             images = images.to(device=device, dtype=torch.float32)
-            batch_labels = batch_labels.to(device=device, dtype=torch.float32)
+            batch_labels = batch_labels.to(device=device)
 
             optimizer.zero_grad()
